cleaned up some issues created by last merge
