+++ Cargo.toml
@@ -17,7 +17,6 @@ members = ["crates/coordinator/holons",
 holochain_integrity_types = "0.2"
 hdi = "0.2"
 hdk = "0.1"
-holochain_integrity_types = "0.2"
 serde = "1"
 derive-new = "0.5"
 thiserror = "1"
@@ -27,7 +26,7 @@ thiserror = "1"
 path = "crates/coordinator/holons"
 
 [workspace.dependencies.holons_integrity]
-path = "crates/integrity/holons"
+path = "crates/integrity/holons_integrity"
 
 [workspace.dependencies.shared_types_holon]
 path = "crates/shared_types/holon"
