+++ crates/coordinator/holons/tests/holon_tests.rs
@@ -31,7 +31,7 @@ use shared_types_holon::holon_node::{PropertyName, PropertyMap, PropertyValue};
 ///
 /// Test Outline:
 /// 1. After initial setup, perform a `get_all_holons`, with an expectation of an empty result
-/// 2. For each test_holon in the `holons` vector, 
+/// 2. For each test_holon in the `holons_integrity` vector,
 ///      * create a new holon (to serve as builder)
 ///      * iterate through the test_holon's properties, invoking external app_property_value for each.
 ///      * commit the holon
@@ -39,10 +39,10 @@ use shared_types_holon::holon_node::{PropertyName, PropertyMap, PropertyValue};
 /// /// 3. Once all data has been created in DHT, perform `get_all_holons` and verify the result.
 ///
 /// Note that this will exercise, create, get, and get_all capabilities across a variety of
-/// holons
+/// holons_integrity
 ///
 /// To selectively run JUST THE TESTS in this file, use:
-///      cargo test -p holons --test holon_tests  -- --show-output
+///      cargo test -p holons_integrity --test holon_tests  -- --show-output
 ///
 #[rstest]
 #[case::create_value_descriptor_holon(new_holons_fixture())]
@@ -78,7 +78,7 @@ async fn rstest_holon_capabilities(
 
     let mut created_action_hashes: Vec<ActionHash> = Vec::new();
 
-    // Iterate through the vector of test holons, building & creating each holon,
+    // Iterate through the vector of test holons_integrity, building & creating each holon,
     // then get the created holon and compare it to the generated descriptor.
     for test_holon in test_holons.clone() {
         let p_count = test_holon.property_map.len();
@@ -147,7 +147,7 @@ async fn rstest_holon_capabilities(
             .await;
     }
 
-    debug!("Performing get_all_holons here to ensure all holons have been deleted.\n");
+    debug!("Performing get_all_holons here to ensure all holons_integrity have been deleted.\n");
 
     let fetched_holons : Vec<Holon> = conductor
         .call(
@@ -158,6 +158,6 @@ async fn rstest_holon_capabilities(
         .await;
 
     assert_eq!(0, fetched_holons.len());
-    println!("...Success! All holons have been deleted. \n");
+    println!("...Success! All holons_integrity have been deleted. \n");
 }
