--- crates/integrity/holons/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "holons_integrity"
-version = "0.0.1"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-name = "holons_integrity"
-
-[dependencies]
-hdi = { workspace = true }
-holochain_integrity_types = { workspace = true }
-
-serde = { workspace = true }
-derive-new = { workspace = true }
-thiserror = { workspace = true }
-
-shared_types_holon = {workspace = true}
-
