+++ crates/integrity/holons_integrity/src/holon_node.rs
