--- crates/integrity/holons_integrity/src/holon_validators.rs
@@ -1,101 +1,0 @@
-use hdi::prelude::*;
-use shared_types_holon::holon::Holon;
-
-pub fn validate_create_holon(
-    _action: EntryCreationAction,
-    _holon: Holon,
-) -> ExternResult<ValidateCallbackResult> {
-    Ok(ValidateCallbackResult::Valid)
-}
-pub fn validate_update_holon(
-    _action: Update,
-    _holon: Holon,
-    _original_action: EntryCreationAction,
-    _original_holon: Holon,
-) -> ExternResult<ValidateCallbackResult> {
-    Ok(ValidateCallbackResult::Valid)
-}
-pub fn validate_delete_holon(
-    _action: Delete,
-    _original_action: EntryCreationAction,
-    _original_holon: Holon,
-) -> ExternResult<ValidateCallbackResult> {
-    Ok(ValidateCallbackResult::Valid)
-}
-pub fn validate_create_link_holon_updates(
-    _action: CreateLink,
-    base_address: AnyLinkableHash,
-    target_address: AnyLinkableHash,
-    _tag: LinkTag,
-) -> ExternResult<ValidateCallbackResult> {
-    let action_hash = ActionHash::from(base_address);
-    let record = must_get_valid_record(action_hash)?;
-    let _holon: crate::Holon = record
-        .entry()
-        .to_app_option()
-        .map_err(|e| wasm_error!(e))?
-        .ok_or(
-            wasm_error!(
-                WasmErrorInner::Guest(String::from("Linked action must reference an entry"))
-            ),
-        )?;
-    let action_hash = ActionHash::from(target_address);
-    let record = must_get_valid_record(action_hash)?;
-    let _holon: crate::Holon = record
-        .entry()
-        .to_app_option()
-        .map_err(|e| wasm_error!(e))?
-        .ok_or(
-            wasm_error!(
-                WasmErrorInner::Guest(String::from("Linked action must reference an entry"))
-            ),
-        )?;
-    Ok(ValidateCallbackResult::Valid)
-}
-pub fn validate_delete_link_holon_updates(
-    _action: DeleteLink,
-    _original_action: CreateLink,
-    _base: AnyLinkableHash,
-    _target: AnyLinkableHash,
-    _tag: LinkTag,
-) -> ExternResult<ValidateCallbackResult> {
-    Ok(
-        ValidateCallbackResult::Invalid(
-            String::from("HolonUpdates links cannot be deleted"),
-        ),
-    )
-}
-pub fn validate_create_link_all_holons(
-    _action: CreateLink,
-    _base_address: AnyLinkableHash,
-    target_address: AnyLinkableHash,
-    _tag: LinkTag,
-) -> ExternResult<ValidateCallbackResult> {
-    // Check the entry type for the given action hash
-    let action_hash = ActionHash::from(target_address);
-    let record = must_get_valid_record(action_hash)?;
-    let _holon: crate::Holon = record
-        .entry()
-        .to_app_option()
-        .map_err(|e| wasm_error!(e))?
-        .ok_or(
-            wasm_error!(
-                WasmErrorInner::Guest(String::from("Linked action must reference an entry"))
-            ),
-        )?;
-    // TODO: add the appropriate validation rules
-    Ok(ValidateCallbackResult::Valid)
-}
-pub fn validate_delete_link_all_holons(
-    _action: DeleteLink,
-    _original_action: CreateLink,
-    _base: AnyLinkableHash,
-    _target: AnyLinkableHash,
-    _tag: LinkTag,
-) -> ExternResult<ValidateCallbackResult> {
-    Ok(
-        ValidateCallbackResult::Invalid(
-            String::from("AllHolons links cannot be deleted"),
-        ),
-    )
-}
