+++ crates/integrity/holons_integrity/src/smartlink.rs
